Increase fastcgi_buffers size
The fastcgi_buffers size defines the buffer used when reading the response from the
proxied server. A too low value causes many requests to buffer to disk, that
could have been served from memory instead.

It's easy to notice this in the nginx logs by the following warning:
    an upstream response is buffered to a temporary file while reading upstream

Most requests can now be served without buffering to a temporary file.